Basically useless site, but now with Templates!
I think Django's template system is dumb. Or at least the way they explain it in the tutorial is dumb.
By default they want you to use an absolute path, but basically everything will eventually get deployed somewhere, so the path will change. And Django seems to be setup to compartmentalize everything into 'apps' so I expected that they would have the templates under the 'app' directory. Most people will want cross-app templates to give the site a common styling (layouts in Rails), but I was expecting that to be handled by an import system. We'll see if this makes more sense as I dig deeper.

The path solution for the templates comes from StackOverflow.