Trying to get images with the Twitter "media" entity.
In the first page of results at the time I was searching it didn't return anything with a "media" entity. Entities are parsed out bits of the tweet like urls, hashtags and images. The "media" entity only works for images uploaded through twitter's image uploader and not things like Instagram. So apparently there aren't a lot of people using Twitter's own upload facility and I'll have to do more complex parsing.

I think my goal now is going to be to create a twitter-fall of cute animal images because the internet doesn't have enough of that. I was hoping it would be simple and I could get enough results with the media entities, but I'll have to parse everything out myself.

The embeded images Twitter should be able to tell me exactly how to get the image itself without parsing the whole Instagram or Twitpic page. For example, from the regular Instagram url:
  http://instagram.com/p/M8GM4ayXgR/

I can just add /media/?size=l to get just the image:
  http://instagram.com/p/M8GM4ayXgR/media/?size=l

Some of the others aren't so easy and will require parsing thing instead of just tacking on params. And Twitter itself uses Plixi to get stuff out of lockerz.com.